Fix approach to avoid infinite spinning
No longer locks steering directly to the relative velocity vector.
When getting close to this target the ship would start spinning
around it, sometimes not recovering.

Instead, the steering is locked to steer_lock and this is updated
to point to the target vector only when safe to do so.

Added some better vector math to break out of loops when the vector
flips direction, indicating that the ship has overshot the target.